requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
build==1.2.2
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
constantly==23.10.4
crashtest==0.4.1
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.1
dnspython==2.6.1
docker==7.1.0
dulwich==0.21.7
Faker==28.4.1
fastjsonschema==2.20.0
filelock==3.16.0
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
installer==0.7.0
itemadapter==0.9.0
itemloaders==1.3.1
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==24.3.1
lxml==5.3.0
MarkupSafe==2.1.5
mongoengine==0.29.0
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
parsel==1.9.1
pexpect==4.9.0
pika==1.3.2
pkginfo==1.11.1
platformdirs==4.3.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
Protego==0.3.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pymongo==4.8.0
pyOpenSSL==24.2.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.7.0
rapidfuzz==3.9.7
redis==5.0.8
redis-lru==0.1.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sql==2022.4.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
tldextract==5.1.2
tomlkit==0.13.2
trove-classifiers==2024.7.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.4
w3lib==2.2.1
xattr==1.1.0
yarl==1.11.1
zope.interface==7.0.3
